SAVANEX Project/backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.7.0
reportlab==4.2.0
django-extensions==3.2.3
coverage==7.5.0
pytest-django==4.8.0
